CREST-133: Migrate CREST from Futures to Promises OPENDJ-1285: Migrate SDK from Futures to Promises
Initial commit of Promises API including the Promise interface, a basic Promise implementation (PromiseImpl), a utility class for manipulating Promises, as well as various call-back interfaces. The API has been aligned with existing CREST interfaces, especially ResultHandler and Future.

PromiseImpl uses intrinsic lock/monitor for synchronization which results in a more compact representation compared with an AbstractQueuedSynchronizer based implementation. Intrinsic locking should be ok in this case, because the blocking Promise/Future methods (getXXX) are not usually used and, when they are, there should rarely be any contention. Having said that, the performance still needs to be evaluated.

TODO:

* validate Closure (JDK8) support
* evaluate performance
* add support for progress notification, e.g. ProgressiblePromise (name TBD)
* unit tests once APIs are more stable.