http://sources.forgerock.org/cru/FR-62
A few random fixes - 

Router.js checks to make sure your URL args are strings, rather than simply not "undefined" - this was needed
because the "url" handlebars helper was sometimes passing in an object rather than an array of strings as the arguments. 
In this case, we want to use the defaults.

forms.less adds a max-width for group-validation errors, to prevent issues with layout when rendering large validation messages.

UserValidators now accepts a wider format of "reauthentication" calls, so that any successful response will be assumed to be valid,
and any unsuccessful one (non 200) will be assumed to be invalid.