Fix for DOCS-76: Cannot copy/paste examples from PDF
Tested with forgerock-doc-maven-plugin-test

In contrast to the description for this issue, the fix does not actually fix the examples in PDF formats. Instead, it allows authors to add continuation characters in <screen> examples, so that they can be copy/pasted from documents (in HTML, PDF, or other formats). Previously the continuation characters clashed with JavaScript line folding/flattening in HTML.