Removed an invalid test case. Thanks Peter for pointing out this problem.
Its purpose was to verify that TimeService.SYSTEM.now() was really returning the now instant by comparing the returned value with values capture just before and after (checking that now() was in between).

But it actually turns out that System.currentTimeMillis() can effectively goes backward ([1]) so I cannot assume that the returned values are monotonic. That makes this test irrelevant.
 I couldn't find any other way to test the return result of TimeService.SYSTEM.now().

 [1]. http://stackoverflow.com/questions/2978598/will-sytem-currenttimemillis-always-return-a-value-previous-calls