CR-4279 - ldapjs fork from https://github.com/mcavage/node-ldapjs/tree/master/lib/filters
This library code was originally created to work as part of an LDAP server for NodeJS. Part of this LDAP server code was some javascript that was used to parse and handle LDAP query strings. It has been modified for use in the browser (AMD modules instead of commonjs, support for Internet Explorer-style JS, etc...).