Fix for DOCS-180: Move admonition graphics to common content
Tested with forgerock-doc-maven-plugin-test

It's possible that something else went wrong when Gene was testing this method originally with DOCS-137, because it works just fine when the 3.0.0-SNAPSHOT is picked up during the build.

At present, this fix only affects PDF. I'm going to re-open DOCS-137 to suggest that it be fixed at least for HTML as well.