CR-5817 DOCS-188 Remove JavaScript icons [+][-] for (un)folding <screen>
Some time ago, we were including <screen> listings
without continuation characters for long lines,
making it hard to copy & paste.

So we added a JavaScript function linked to an icon in the HTML
that would unfold <screen> content, making it easier to copy & paste. 

But a better solution was to add continuation characters for long lines,
which we have done throughout the docs since then.
Formatting no longer needs to change in order to copy & paste. 

As attaching this functionality to an icon added by JavaScript
not only is no longer necessary 
but also probably does not help accessibility,
this patch removes the functionality. 

As a side effect, this should fix DOCS-150 as well.