CR-6626 DOCS-278 Do not create extra build directories just for Olink DBs
This patch prevents the plugin from creating content under target/<format>,
where format is one of the formats to be generated with docbkx-tools,
even if the build directories are elsewhere.
Although not a problem for the final output, it can be confusing to writers
who are looking for the docs they just built in the pre-site phase.

Tested with forgerock-doc-maven-plugin-test (where project.build.directory is build)
and all the olinks still function.

Thanks to David G. for the issue and for review.