Quick fix : switch back to assertj 1.6.1
The upgrade to assertj 1.7.1 had some side effects on the downstream
projects : compilation failures, forced to upgrade to 1.7.1.
While staying on 1.6.1, that removes these failures.

On the other side, staying on 1.6.1, with the API changes, raises a
javac bug in 1.7.x (a StackOverFlowError). The only way to work around
that bug is to use the ECJ compiler.

So this commit is mainly to get the situation back to normal and see how
we plan the upgrade to 1.7.1 (and later 2.0.0).