CMN-107 Update description of version numbers
This patch updates the shared documentation about versioning
to say that trailing .0s in version numbers are optional.

Projects do not have to drop trailing .0s in the documentation
or anywhere else.
If a project does however drop a trailing .0,
this shared documentation will be in accord with that change.